fixed it by adding +1 when going down each child, so it can correctly count the height in steps

+++ 14-binary_tree_balance.c
@@ -1,9 +1,9 @@
 #include "binary_trees.h"
 
 /**
- * binary_tree_height - helper to get the height of a tree
- * @tree: pointer to node
- * Return: height or 0
+ * binary_tree_height - returns the height of a binary tree
+ * @tree: pointer to the root node of the subtree
+ * Return: height (number of steps to the deepest leaf)
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
@@ -14,6 +14,7 @@ size_t binary_tree_height(const binary_tree_t *tree)
 
 	if (tree->left)
 		left = 1 + binary_tree_height(tree->left);
+
 	if (tree->right)
 		right = 1 + binary_tree_height(tree->right);
 
@@ -21,14 +22,15 @@ size_t binary_tree_height(const binary_tree_t *tree)
 }
 
 /**
- * binary_tree_balance - measures balance factor of a node
- * @tree: pointer to the node
- * Return: balance (left height - right height)
+ * binary_tree_balance - gets the balance factor of a binary tree node
+ * @tree: pointer to the node to check
+ * Return: balance factor = height of left - height of right
  */
+
 int binary_tree_balance(const binary_tree_t *tree)
 {
 	if (tree == NULL)
 		return (0);
 
-	return (binary_tree_height(tree->left) - binary_tree_height(tree->right));
+	return ((int)binary_tree_height(tree->left) - (int)binary_tree_height(tree->right));
 }
